updating recycle view using data from firebase

+++ app/src/main/java/com/dmytroandriichuk/cmediacal/ui/search/model/SearchListChildItem.kt
@@ -3,7 +3,7 @@ package com.dmytroandriichuk.cmediacal.ui.search.model
 //POJO item of SearchListChildAdapter
 data class SearchListChildItem(
         val name: String,
-        val price: Float
+        val price: Double
 ) {
     override fun toString(): String {
         return "SearchListChildItem(name='$name', price=$price)"
